createまでは、作成

+++ backend/contracts/Wallet/WalletFactoryV2.sol
@@ -1,22 +1,26 @@
 // SPDX-License-Identifier: MIT
 pragma solidity >=0.8.9;
 
-import './MultiSigWallet.sol';
-import './../Token/MyToken.sol';
+import "./MultiSigWalletV2.sol";
+import "./../Token/MyToken.sol";
 
 /**
  * WalletFactoryコントラクト
  */
 contract WalletFactoryV2 {
-
     // MultiSigWallet型の配列
     MultiSigWallet[] public wallets;
     // 関数から返すことのできる最大値
     uint256 constant maxLimit = 20;
 
     // インスタンスが生成された時のイベント
-    event WalletCreated (MultiSigWallet indexed wallet, string name, address[] owners, uint required);
-    event transferIDQ (address to, uint256 amount);
+    event WalletCreated(
+        MultiSigWallet indexed wallet,
+        string name,
+        address[] owners,
+        uint256 required
+    );
+    event transferIDQ(address to, uint256 amount);
 
     /**
      * MultiSigWalletのインスタンス数を取得する関数
@@ -31,10 +35,10 @@ contract WalletFactoryV2 {
      * @param _owners アドレスの配列
      * @param _required 閾値
      */
-    function createWallet (
-        string memory _name, 
-        address[] memory _owners, 
-        uint _required
+    function createWallet(
+        string memory _name,
+        address[] memory _owners,
+        uint256 _required
     ) public {
         // インスタンスを生成
         MultiSigWallet wallet = new MultiSigWallet(_name, _owners, _required);
@@ -47,9 +51,12 @@ contract WalletFactoryV2 {
     /**
      * 作成済みウォレットの情報を取得するメソッド
      */
-    function getWallets(uint256 limit, uint256 offset) public view returns (MultiSigWallet[] memory coll) {
-        
-        require (offset <= walletsCount(), "offset out of bounds");
+    function getWallets(uint256 limit, uint256 offset)
+        public
+        view
+        returns (MultiSigWallet[] memory coll)
+    {
+        require(offset <= walletsCount(), "offset out of bounds");
         // 最大値を上回っている場合は、limitを格納する。
         uint256 size = walletsCount() - offset;
         size = size < limit ? size : limit;
@@ -61,7 +68,7 @@ contract WalletFactoryV2 {
             coll[i] = wallets[offset + i];
         }
 
-        return coll;    
+        return coll;
     }
 
     /**
@@ -71,9 +78,12 @@ contract WalletFactoryV2 {
         // create IDQToken object
         MyToken token = MyToken(_tokenAddr);
         // check
-        require(token.balanceOf(address(this)) >= _amount, "insufficient Tokens");
+        require(
+            token.balanceOf(address(this)) >= _amount,
+            "insufficient Tokens"
+        );
         // approve
-        token.approve(address(this), _amount); 
+        token.approve(address(this), _amount);
         // transfer
         token.transfer(msg.sender, _amount);
 
@@ -85,4 +95,3 @@ contract WalletFactoryV2 {
      */
     receive() external payable {}
 }
-
